Making s3 deploy less redundant

+++ s3-tarball-deploy/recipes/default.rb
@@ -26,25 +26,26 @@ node[:s3_tarball_deploy].each do |name, config|
     recursive true
   end
 
-  bash "deploy" do
-    user "root"
-    cwd "/tmp"
-    code <<-EOH
-    tar -xzf #{config[:file]}
-    mv #{config[:file]} #{config[:deploy_parent]}/packages/#{config[:deploy_target]}.tar.gz
-    mv /tmp/#{config[:extracted_folder]} #{config[:deploy_parent]}/releases/#{config[:deploy_target]}
-    chown #{config[:owner]}:#{config[:group]} -R #{config[:deploy_parent]}/releases/#{config[:deploy_target]}
-    chmod 775 -R #{config[:deploy_parent]}/releases/#{config[:deploy_target]}
-    EOH
-    not_if do File.exists?(config[:deploy_parent] + "/releases/" + config[:deploy_target]) end
-  end
+ if not File.exists?(config[:deploy_parent] + "/releases/" + config[:deploy_target]) then 
+    bash "deploy" do
+      user "root"
+      cwd "/tmp"
+      code <<-EOH
+      tar -xzf #{config[:file]}
+      mv #{config[:file]} #{config[:deploy_parent]}/packages/#{config[:deploy_target]}.tar.gz
+      mv /tmp/#{config[:extracted_folder]} #{config[:deploy_parent]}/releases/#{config[:deploy_target]}
+      chown #{config[:owner]}:#{config[:group]} -R #{config[:deploy_parent]}/releases/#{config[:deploy_target]}
+      chmod 775 -R #{config[:deploy_parent]}/releases/#{config[:deploy_target]}
+      EOH
+    end
 
-  link config[:deploy_parent] + "/releases/" + config[:symlink] do
-    to config[:deploy_parent] + "/releases/" + config[:deploy_target]
-    not_if do File.exists?(config[:deploy_parent] + "/releases/" + config[:symlink]) end
-  end
+    link config[:deploy_parent] + "/releases/" + config[:symlink] do
+      to config[:deploy_parent] + "/releases/" + config[:deploy_target]
+      not_if do File.exists?(config[:deploy_parent] + "/releases/" + config[:symlink]) end
+    end
 
-  file "/tmp/#{config[:file]}" do 
-    action :delete
+    file "/tmp/#{config[:file]}" do 
+      action :delete
+    end
   end
 end
